+++ apps/web/src/components/Cards/SurfaceCards/BlobCard.tsx
@@ -1,22 +1,20 @@
 import type { FC } from "react";
 
-import { RollupIcon } from "~/components/RollupIcon";
+import { RollupBadge } from "~/components/Badges/RollupBadge";
+import { StorageBadge } from "~/components/Badges/StorageBadge";
 import { Skeleton } from "~/components/Skeleton";
-import { StorageIcon } from "~/components/StorageIcon";
 import { useBreakpoint } from "~/hooks/useBreakpoint";
-import type { Rollup } from "~/types";
+import type { Blob, Rollup } from "~/types";
 import { buildBlobRoute, formatBytes } from "~/utils";
-import type { DeserializedBlob } from "~/utils";
 import { Link } from "../../Link";
 import { CardField } from "../Card";
 import { SurfaceCardBase } from "./SurfaceCardBase";
 
 type BlobCardProps = Partial<{
   blob: Pick<
-    DeserializedBlob,
+    Blob,
     "versionedHash" | "commitment" | "size" | "dataStorageReferences" | "proof"
   > & { rollup?: Rollup | null };
-  rollup: Rollup | null;
   compact?: boolean;
   className?: string;
 }>;
@@ -51,7 +49,9 @@ const BlobCard: FC<BlobCardProps> = ({
               </span>
               <Link href={buildBlobRoute(versionedHash)}>{versionedHash}</Link>
             </div>
-            {rollup ? <RollupIcon key={rollup} rollup={rollup} /> : null}
+            {rollup ? (
+              <RollupBadge key={rollup} rollup={rollup} compact />
+            ) : null}
           </div>
         ) : (
           <Skeleton width={isCompact ? undefined : 630} />
@@ -75,11 +75,12 @@ const BlobCard: FC<BlobCardProps> = ({
               <span>·</span>
               <div className="flex flex-row gap-1">
                 {dataStorageReferences.map(({ storage, url }) => (
-                  <StorageIcon
+                  <StorageBadge
                     key={storage}
                     storage={storage}
                     url={url}
                     size="md"
+                    compact
                   />
                 ))}
               </div>