+++ apps/web/src/types/entities.ts
@@ -1,16 +1,11 @@
+import type { RouterOutputs } from "@blobscan/api";
 import type {
   BlobStorage as BlobStorageEnum,
   Rollup as RollupEnum,
   Category as CategoryEnum,
 } from "@blobscan/api/enums";
 
-import type { MakeFieldRequired } from "./helpers";
-import type {
-  GetAllBlobsOutput,
-  GetByBlobIdOutput,
-  GetByBlockIdOutput,
-  GetTxByHashOutput,
-} from "./routers";
+import type { EChartCompliant, MakeFieldRequired } from "./helpers";
 
 export type BlobStorage = Lowercase<BlobStorageEnum>;
 
@@ -18,17 +13,18 @@ export type Rollup = Lowercase<RollupEnum>;
 
 export type Category = Lowercase<CategoryEnum>;
 
-export type Blob = GetByBlobIdOutput;
+export type Blob = RouterOutputs["blob"]["getByBlobId"];
 
-export type BlobOnTransaction = GetAllBlobsOutput["blobs"][number];
+export type BlobOnTransaction =
+  RouterOutputs["blob"]["getAll"]["blobs"][number];
 
-export type Block = GetByBlockIdOutput;
+export type Block = RouterOutputs["block"]["getByBlockId"];
 
-export type Transaction = GetTxByHashOutput;
+export type Transaction = RouterOutputs["tx"]["getByHash"];
 
 export type TransactionWithExpandedBlock = MakeFieldRequired<
   Transaction,
-  "block" | "blobGasBaseFee"
+  "block"
 >;
 
 export type TransactionWithExpandedBlob = MakeFieldRequired<
@@ -38,7 +34,7 @@ export type TransactionWithExpandedBlob = MakeFieldRequired<
 
 export type TransactionWithExpandedBlockAndBlob = MakeFieldRequired<
   Transaction,
-  "blobs" | "block" | "blobGasBaseFee"
+  "blobs" | "block"
 >;
 
 type BlockExpandedTransactionWithExpandedBlobs = MakeFieldRequired<
@@ -62,3 +58,13 @@ export type BlobWithExpandedTransaction = MakeFieldRequired<
   BlobOnTransaction,
   "transaction"
 >;
+
+export type DailyStats = RouterOutputs["stats"]["getDailyStats"][number];
+
+export type OverallStats = RouterOutputs["stats"]["getOverallStats"];
+
+export type EChartCompliantDailyStats = EChartCompliant<DailyStats>;
+
+export type EChartCompliantOverallStats = EChartCompliant<OverallStats>;
+
+export type RollupDailyStats = RouterOutputs["stats"]["getRollupDailyStats"];