+++ packages/api/test/tx.test.ts
@@ -78,17 +78,15 @@ describe("Transaction router", async () => {
   });
 
   describe("getByHash", () => {
-    runExpandsTestsSuite(
-      "transaction",
-      ["block", "blob", "blob_data"],
-      (expandsInput) =>
-        caller.tx.getByHash({ hash: "txHash001", ...expandsInput })
+    runExpandsTestsSuite("transaction", ["block", "blob"], (expandsInput) =>
+      caller.tx.getByHash({ hash: "txHash001", ...expandsInput })
     );
 
     it("should get a transaction by hash correctly", async () => {
       const result = await caller.tx.getByHash({
         hash: "txHash001",
       });
+
       expect(result).toMatchSnapshot();
     });