+++ packages/env/index.ts
@@ -1,16 +1,35 @@
 import type { RefinementCtx } from "@blobscan/zod";
 import {
   z,
-  blobStorageSchema,
   booleanSchema,
   createEnv,
   presetEnvOptions,
-  nodeEnvSchema,
   maskPassword,
-  networkSchema,
   maskSensitiveData,
-  prismaBatchOperationsMaxSizeSchema,
 } from "@blobscan/zod";
+
+const nodeEnvSchema = z.enum(["development", "test", "production"]);
+
+const blobStorageSchema = z.enum([
+  "FILE_SYSTEM",
+  "GOOGLE",
+  "POSTGRES",
+  "SWARM",
+] as const);
+
+const networkSchema = z.enum([
+  "mainnet",
+  "holesky",
+  "sepolia",
+  "gnosis",
+  "chiado",
+  "devnet",
+]);
+
+const prismaBatchOperationsMaxSizeSchema = z.coerce
+  .number()
+  .positive()
+  .default(100_000);
 
 function requireIfEnvEnabled(envName: string) {
   return (value: unknown, ctx: RefinementCtx) => {
@@ -30,6 +49,8 @@ export const env = createEnv({
       BLOBSCAN_API_PORT: z.coerce.number().positive().default(3001),
       NODE_ENV: nodeEnvSchema.optional(),
       SECRET_KEY: z.string(),
+      BLOB_DATA_API_ENABLED: booleanSchema.default("true"),
+      BLOB_DATA_API_KEY: z.string().optional(),
       CHAIN_ID: z.coerce.number().positive().default(1),
       DENCUN_FORK_SLOT: z.coerce.number().optional(),
       NETWORK_NAME: networkSchema.default("mainnet"),
@@ -83,19 +104,25 @@ export const env = createEnv({
         .url()
         .default("http://localhost:4318"),
 
-      // ETH Price (default: every hour)
-      ETH_PRICE_SYNCER_CRON_PATTERN: z
-        .enum(["0 * * * *", "0 0 * * *", "* * * * *"])
-        .default("0 * * * *"),
-      ETH_PRICE_SYNCER_CHAIN_ID: z.coerce.number().optional(),
+      /*
+       * ====================
+       *  ETH/USD Price Feed
+       * ====================
+       *
+       * https://polygonscan.com/address/0xF9680D99D6C9589e2a93a78A04A279e509205945
+       * https://data.chain.link/
+       */
+
+      // ETH Price is retrieved every second from the Chainlink: ETH/USD oracle
+      // in the Polygon network.
+      ETH_PRICE_SYNCER_ENABLED: booleanSchema.default("false"),
+      ETH_PRICE_SYNCER_CRON_PATTERN: z.string().default("* * * * *"),
+      ETH_PRICE_SYNCER_CHAIN_ID: z.coerce.number().default(137),
       ETH_PRICE_SYNCER_CHAIN_JSON_RPC_URL: z.string().url().optional(),
       ETH_PRICE_SYNCER_ETH_USD_PRICE_FEED_CONTRACT_ADDRESS: z
         .string()
-        .optional(),
-      ETH_PRICE_SYNCER_TIME_TOLERANCE: z.coerce
-        .number()
-        .positive()
-        .default(3600),
+        .default("0xF9680D99D6C9589e2a93a78A04A279e509205945"),
+      ETH_PRICE_SYNCER_TIME_TOLERANCE: z.coerce.number().positive().optional(),
 
       /*
        * =====================