+++ packages/syncers/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @blobscan/syncers
 
+## 0.4.2
+
+### Patch Changes
+
+- [#789](https://github.com/Blobscan/blobscan/pull/789) [`3b54479`](https://github.com/Blobscan/blobscan/commit/3b54479da9d7e49398bd8f2b6ad8882b9fcb7a24) Thanks [@PJColombo](https://github.com/PJColombo)! - Resolved issue where the ETH-USD price feed syncer was creating new price feed state rows in the database on every run
+
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies [[`186aa87`](https://github.com/Blobscan/blobscan/commit/186aa873c0970d9717cb04db1a6ad4adeb1faf92)]:
+  - @blobscan/db@0.17.0
+
+## 0.4.0
+
+### Minor Changes
+
+- [#731](https://github.com/Blobscan/blobscan/pull/731) [`29b205a`](https://github.com/Blobscan/blobscan/commit/29b205a0e096a7cf42f26554c2f4818c94303295) Thanks [@luis-herasme](https://github.com/luis-herasme)! - Added ETH price syncer
+
+### Patch Changes
+
+- Updated dependencies [[`29b205a`](https://github.com/Blobscan/blobscan/commit/29b205a0e096a7cf42f26554c2f4818c94303295)]:
+  - @blobscan/db@0.16.0
+
 ## 0.3.5
 
 ### Patch Changes